refactor: modularização e empacotamento
