+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "client"
+name = "fishnet"
 version = "0.1.0"
 edition = "2024"
