+++ src/ui/mod.rs
@@ -1,0 +1,12 @@
+
+
+mod cli;
+mod fisher;
+mod tui;
+
+pub use cli::Args;
+
+pub use fisher::FishCatalog;
+pub use fisher::fishing;
+
+pub use tui::eval;
